+++ src/components/charts/ColumnChartComponent.vue
