+++ src/components/charts/HorizontalBarChartComponent.vue
