+++ src/components/charts/VerticalBarChartComponent.vue
