+++ src/components/filters/CheckboxFilterComponent.vue
